# Copy over TerminalWriter from py, simplify and optimize it