--- a/src/_pytest/_io/__init__.py
+++ b/src/_pytest/_io/__init__.py
@@ -1,6 +1,8 @@
+from .terminalwriter import get_terminal_width
 from .terminalwriter import TerminalWriter
 
 
 __all__ = [
     "TerminalWriter",
+    "get_terminal_width",
 ]