… changelog/7135.breaking.rst
… src/_pytest/config/argparsing.py
… src/_pytest/pastebin.py
… src/_pytest/python.py
… src/_pytest/reports.py
… src/_pytest/runner.py
… src/_pytest/setuponly.py
… src/_pytest/terminal.py
… testing/code/test_excinfo.py
… testing/code/test_terminal_writer.py
… testing/io/test_terminalwriter.py
… testing/test_config.py