I was looking at some profiles, and noticed that when there is a lot of output, `TerminalWriter` features prominently. This prompted me to look into it.

`TerminalWriter` comes from `py`. `py` is frozen and needs to maintain backward compat and generality, so it's hard to do anything with that. So I decided to copy it over to pytest where it can be modified more freely. I believe pytest did something similar with the capturing infrastructure back in the day. And this also coincides with the goal of eventually removing the `py` dependency.

The commits in this PR detail every change I made starting from a straight copy. The main simplifications are:

- Remove all the Windows specific stuff -- my understanding is that coloroma already takes care of it, but I don't have Windows to test with.
- Remove functions unused by pytest.
- Remove Python2 stuff.
- Remove support for writing `bytes` which pytest doesn't need.
- Some general cleanups.

I also merged the changes from our custom `TerminalWriter` subclass, and ported the tests from py.

The file in `py` had 421, after all changes it is 180 lines not including the code from the subclass.

The optimizations are:

- Reduced overhead due to simplifications.
- Reduce amount of flushes to underlying file.
- Avoid expensive work of computing line widths when possible.

---

For a benchmark I used a semi-realistic scenario -- not a a micro-benchmark specifically of TerminalWriter, but something that does print a lot. Running the following with `pytest --collect-only`:

```py
import pytest
@pytest.mark.parametrize("x", range(50000))
def test_foo(x): pass
```

Before (leaving only relevant entries):

```
============================ no tests ran in 12.39s ============================
         23778937 function calls (22924066 primitive calls) in 12.674 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    50008    0.099    0.000    2.727    0.000 terminalwriter.py:270(line)
   100016    0.141    0.000    2.587    0.000 terminalwriter.py:245(write)
   100015    0.203    0.000    1.812    0.000 terminalwriter.py:258(_update_chars_on_current_line)
   100015    0.113    0.000    1.518    0.000 terminalwriter.py:72(get_line_width)
  1489288    0.710    0.000    1.125    0.000 terminalwriter.py:74(<genexpr>)
   100015    0.085    0.000    0.600    0.000 terminalwriter.py:399(write_out)
   100015    0.496    0.000    0.496    0.000 {method 'flush' of '_io.TextIOWrapper' objects}
```

After:

```
============================ no tests ran in 9.59s =============================
         17981953 function calls (17128590 primitive calls) in 9.741 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    50007    0.042    0.000    0.176    0.000 terminalwriter.py:175(line)
   100014    0.090    0.000    0.134    0.000 terminalwriter.py:159(write)
```

So at least in this scenario the overhead due to `TerminalWriter` is mostly eliminated.