@bluetech 2020-04-30
Thanks for the review @RonnyPfannschmidt!

I made the changes you requested, added a few small cleanups I noticed, and added a changelog detailing the breaking changes.

> we should probably use `PYTEST_COLORS` here

Unlike the other changes which *might* break some plugins, this one would be a breaking change for users. If we do decide to add a `PYTEST_COLORS`, it could subsume `PY_COLORS`, but would need to be documented properly. I'd be happy to do it, but prefer to do it in a separate PR.