@RonnyPfannschmidt 2020-04-30
Good call, this needs some work for configuration migration wrt py_colors vs a new name