@nicoddemus 2020-05-02
> as a followup i would like to note down an investigation for switching the kwarg based markup details with something more fluent

Definitely, that's something I've been thinking lately, after introducing Syntax Highlighting.

I've seen somewhere (I think in a `ward` issue/discussion) about [rich](https://pypi.org/project/rich). It supports inline markup. We could consider using the library directly or as inspiration.

But of course this is off-topic here. 😁