@nicoddemus 2020-05-02
I've tested this on Windows and didn't encounter any issues.

Great work @bluetech!