@RonnyPfannschmidt 2020-05-02
@nicoddemus a few, years ago I tried to do terminalwriter in terms of click